Move rpc masternode methods into separate file
