+++ divi/src/MasternodeModule.cpp
@@ -31,18 +31,6 @@
 #include <ThreadManagementHelpers.h>
 
 
-#define MN_WINNER_MINIMUM_AGE 8000    // Age in seconds. This should be > MASTERNODE_REMOVAL_SECONDS to avoid misconfigured new nodes in the list.
-
-template <typename T>
-static T readFromHex(std::string hexString)
-{
-    std::vector<unsigned char> hex = ParseHex(hexString);
-    CDataStream ss(hex,SER_NETWORK,PROTOCOL_VERSION);
-    T object;
-    ss >> object;
-    return object;
-}
-
 bool fMasterNode = false;
 bool fLiteMode = false;
 
@@ -103,6 +91,10 @@ CMasternodeSync& MasternodeModule::getMasternodeSynchronization() const
 {
     return *masternodeSync_;
 }
+ bool MasternodeModule::localNodeIsAMasternode() const
+ {
+     return fMasterNode_;
+ }
 
 MasternodeModule mnModule;
 MasternodeNetworkMessageManager& networkMessageManager = mnModule.getNetworkMessageManager();
@@ -216,233 +208,6 @@ void DumpMasternodeDataToDisk()
         flatdb1.Dump(networkMessageManager);
         CFlatDB<MasternodePaymentData> flatdb2("mnpayments.dat", "magicMasternodePaymentsCache");
         flatdb2.Dump(masternodePaymentData);
-    }
-}
-
-void CountNetworks(int& ipv4, int& ipv6, int& onion)
-{
-    int protocolVersion = ActiveProtocol();
-
-    mnodeman.Check();
-    BOOST_FOREACH (CMasternode& mn, networkMessageManager.masternodes)
-    {
-        std::string strHost;
-        int port;
-        SplitHostPort(mn.addr.ToString(), port, strHost);
-        CNetAddr node = CNetAddr(strHost, false);
-        int nNetwork = node.GetNetwork();
-        switch (nNetwork) {
-            case 1 :
-                ipv4++;
-                break;
-            case 2 :
-                ipv6++;
-                break;
-            case 3 :
-                onion++;
-                break;
-        }
-    }
-}
-
-int StableMasternodeCount()
-{
-    LOCK(networkMessageManager.cs);
-    int nStable_size = 0;
-    int nMinProtocol = ActiveProtocol();
-    int64_t nMasternode_Min_Age = MN_WINNER_MINIMUM_AGE;
-    int64_t nMasternode_Age = 0;
-
-    mnodeman.Check();
-    for (auto& mn : networkMessageManager.masternodes) {
-        if (mn.protocolVersion < nMinProtocol) {
-            continue; // Skip obsolete versions
-        }
-        if (sporkManager.IsSporkActive (SPORK_8_MASTERNODE_PAYMENT_ENFORCEMENT)) {
-            nMasternode_Age = GetAdjustedTime() - mn.sigTime;
-            if ((nMasternode_Age) < nMasternode_Min_Age) {
-                continue; // Skip masternodes younger than (default) 8000 sec (MUST be > MASTERNODE_REMOVAL_SECONDS)
-            }
-        }
-        if (!mn.IsEnabled ())
-            continue; // Skip not-enabled masternodes
-
-        nStable_size++;
-    }
-
-    return nStable_size;
-}
-
-MasternodeCountData::MasternodeCountData(
-    ): total(0)
-    , stable(0)
-    , enabledAndActive(0)
-    , enabled(0)
-    , queueCount(0)
-    , ipv4(0)
-    , ipv6(0)
-    , onion(0)
-{
-}
-
-MasternodeCountData GetMasternodeCounts(const CBlockIndex* chainTip)
-{
-    MasternodeCountData data;
-    if (chainTip != nullptr)
-        data.queueCount = masternodePayments.GetMasternodePaymentQueue(chainTip, 0).size();
-
-    CountNetworks(data.ipv4, data.ipv6, data.onion);
-    data.total = networkMessageManager.masternodeCount();
-    data.stable = StableMasternodeCount();
-    data.enabledAndActive = mnodeman.CountEnabled();
-    data.enabled = data.enabledAndActive;
-
-    return data;
-}
-
-std::vector<MasternodeListEntry> GetMasternodeList(std::string strFilter)
-{
-    std::vector<MasternodeListEntry> masternodeList;
-    std::vector<CMasternode> masternodeVector = networkMessageManager.GetFullMasternodeVector();
-    masternodeList.reserve(masternodeVector.size());
-
-    unsigned numberOfBlocksToSearchBackForLastPayment = (masternodeVector.size()*5)/4;
-    for(auto& masternode : masternodeVector)
-    {
-        std::string strVin = masternode.vin.prevout.ToStringShort();
-        std::string strTxHash = masternode.vin.prevout.hash.ToString();
-        uint32_t oIdx = masternode.vin.prevout.n;
-        const std::string collateralAddress = CBitcoinAddress(masternode.pubKeyCollateralAddress.GetID()).ToString();
-        if (strFilter != "" &&
-            strTxHash.find(strFilter) == std::string::npos &&
-            masternode.Status().find(strFilter) == std::string::npos &&
-            collateralAddress.find(strFilter) == std::string::npos)
-        {
-            continue;
-        }
-        masternodeList.emplace_back();
-        MasternodeListEntry& entry = masternodeList.back();
-
-        std::string strStatus = masternode.Status();
-        std::string strHost;
-        int port;
-        SplitHostPort(masternode.addr.ToString(), port, strHost);
-        CNetAddr node = CNetAddr(strHost, false);
-        std::string strNetwork = GetNetworkName(node.GetNetwork());
-
-        entry.network = strNetwork;
-        entry.txHash = strTxHash;
-        entry.outputIndex = oIdx;
-        entry.status = strStatus;
-        entry.collateralAddress = CBitcoinAddress(masternode.pubKeyCollateralAddress.GetID()).ToString();
-        entry.protocolVersion = masternode.protocolVersion;
-        entry.lastSeenTime = (int64_t)masternode.lastPing.sigTime;
-        entry.activeTime = (int64_t)(masternode.lastPing.sigTime - masternode.sigTime);
-        entry.lastPaidTime = (int64_t)masternodePayments.FindLastPayeePaymentTime(masternode,numberOfBlocksToSearchBackForLastPayment);
-        entry.masternodeTier = CMasternode::TierToString(static_cast<MasternodeTier>(masternode.nTier));
-    }
-    return masternodeList;
-}
-
-ActiveMasternodeStatus GetActiveMasternodeStatus()
-{
-    if (!fMasterNode) throw std::runtime_error("This is not a masternode");
-    ActiveMasternodeStatus result;
-
-    CMasternode* pmn = mnodeman.Find(activeMasternode.vin);
-
-    if (pmn) {
-        result.txHash = activeMasternode.vin.prevout.hash.ToString();
-        result.outputIndex = std::to_string(activeMasternode.vin.prevout.n);
-        result.netAddress = activeMasternode.service.ToString();
-        result.collateralAddress = CBitcoinAddress(pmn->pubKeyCollateralAddress.GetID()).ToString();
-        result.statusCode = std::to_string(activeMasternode.status);
-        result.statusMessage = activeMasternode.GetStatus();
-        result.activeMasternodeFound = true;
-        return result;
-    }
-    assert(!result.activeMasternodeFound);
-    return result;
-}
-
-MasternodeStartResult StartMasternode(const CKeyStore& keyStore,std::string alias, bool deferRelay)
-{
-    MasternodeStartResult result;
-    for(const auto& configEntry : masternodeConfig.getEntries())
-    {
-        if(configEntry.getAlias() != alias)
-            continue;
-
-        CMasternodeBroadcast mnb;
-
-        if(!CMasternodeBroadcastFactory::Create(
-                keyStore,
-                configEntry,
-                result.errorMessage,
-                mnb,
-                false,
-                deferRelay))
-        {
-            result.status = false;
-            return result;
-        }
-
-        if (deferRelay)
-        {
-            CDataStream ss(SER_NETWORK,PROTOCOL_VERSION);
-            ss << mnb;
-
-            result.status = true;
-            result.broadcastData = HexStr(ss.str());
-            return result;
-        }
-
-        if(!mnodeman.ProcessBroadcast(activeMasternode, masternodeSync,nullptr, mnb))
-        {
-            LogPrintf("%s - Relaying broadcast vin = %s\n",__func__, mnb.vin);
-            result.status = false;
-            result.errorMessage = "Error processing broadcast";
-            return result;
-        }
-
-        result.status = true;
-        return result;
-    }
-    result.status = false;
-    result.errorMessage = "Invalid alias, couldn't find MN. Check your masternode.conf file";
-    return result;
-}
-
-bool RelayMasternodeBroadcast(std::string hexData, std::string signature)
-{
-    CMasternodeBroadcast mnb = readFromHex<CMasternodeBroadcast>(hexData);
-    if(!signature.empty())
-    {
-        mnb.signature = ParseHex(signature);
-        if(activeMasternode.IsOurBroadcast(mnb,true))
-        {
-            if(activeMasternode.UpdatePing(mnb.lastPing))
-            {
-                LogPrint("masternode","Ping updated successfully!\n");
-            }
-            else
-            {
-                LogPrint("masternode","Ping not updated! Failure to sign!\n");
-            }
-        }
-        else
-        {
-            LogPrint("masternode","This broadcast does not belong to us!\n");
-        }
-    }
-
-    if (mnodeman.ProcessBroadcast(activeMasternode, masternodeSync,nullptr, mnb))
-    {
-        return true;
-    }
-    else
-    {
-        return false;
     }
 }
