+++ divi/src/MasternodeModule.h
@@ -47,7 +47,11 @@ public:
     CActiveMasternode& getActiveMasternode() const;
     CMasternodePayments& getMasternodePayments() const;
     CMasternodeSync& getMasternodeSynchronization() const;
+    bool localNodeIsAMasternode() const;
 };
+
+const CMasternodeSync& GetMasternodeSync();
+CMasternodePayments& GetMasternodePayments();
 
 void ThreadMasternodeBackgroundSync();
 void LockUpMasternodeCollateral(const Settings& settings, std::function<void(const COutPoint&)> walletUtxoLockingFunction);
@@ -60,83 +64,7 @@ bool ShareMasternodeBroadcastWithPeer(CNode* peer,const uint256& inventoryHash);
 bool ShareMasternodePingWithPeer(CNode* peer,const uint256& inventoryHash);
 bool ShareMasternodeWinnerWithPeer(CNode* peer,const uint256& inventoryHash);
 void ForceMasternodeResync();
-const CMasternodeSync& GetMasternodeSync();
-bool RelayMasternodeBroadcast(std::string hexData,std::string signature = "");
-struct MasternodeStartResult
-{
-    bool status;
-    std::string broadcastData;
-    std::string errorMessage;
-    MasternodeStartResult(
-        ): status(false)
-        , broadcastData("")
-        , errorMessage("No Error Data")
-    {}
-};
-MasternodeStartResult StartMasternode(const CKeyStore& keyStore, std::string alias, bool deferRelay);
-struct ActiveMasternodeStatus
-{
-    bool activeMasternodeFound;
-    std::string txHash;
-    std::string outputIndex;
-    std::string netAddress;
-    std::string collateralAddress;
-    std::string statusCode;
-    std::string statusMessage;
-    ActiveMasternodeStatus(
-        ): activeMasternodeFound(false)
-        , txHash()
-        , outputIndex()
-        , netAddress()
-        , collateralAddress()
-        , statusCode()
-        , statusMessage()
-    {}
-};
-ActiveMasternodeStatus GetActiveMasternodeStatus();
-struct MasternodeListEntry
-{
-    std::string network;
-    std::string txHash;
-    uint64_t outputIndex;
-    std::string status;
-    std::string collateralAddress;
-    int protocolVersion;
-    int64_t signatureTime;
-    int64_t lastSeenTime;
-    int64_t activeTime;
-    int64_t lastPaidTime;
-    std::string masternodeTier;
-    MasternodeListEntry(
-        ): network()
-        , txHash()
-        , outputIndex()
-        , status()
-        , collateralAddress()
-        , protocolVersion()
-        , signatureTime()
-        , lastSeenTime()
-        , activeTime()
-        , lastPaidTime()
-        , masternodeTier()
-    {}
-};
-std::vector<MasternodeListEntry> GetMasternodeList(std::string strFilter);
-struct MasternodeCountData
-{
-    int total;
-    int stable;
-    int enabledAndActive;
-    int enabled;
-    int queueCount;
-    int ipv4;
-    int ipv6;
-    int onion;
-    MasternodeCountData();
-};
-MasternodeCountData GetMasternodeCounts(const CBlockIndex* chainTip);
 bool LoadMasternodeDataFromDisk(UIMessenger& uiMessenger,std::string pathToDataDir);
 void DumpMasternodeDataToDisk();
-CMasternodePayments& GetMasternodePayments();
 bool InitializeMasternodeIfRequested(const Settings& settings, bool transactionIndexEnabled, std::string& errorMessage);
 #endif //MASTERNODE_MODULE_H
